Fix tagorder when argument contains a pattern
When argument is a word or contains a pattern,
prefix it with a '/' for the ":tag" and ":stag" commands.

When argument is a pattern starting with '/', don't add the exact match
match anchors '^' and '$'.

:help taglist
taglist({expr}) Returns a list of tags matching the regular expression
{expr}

:help tag-regexp
The ":tag" and ":tselect" commands accept a regular expression argument.
When the argument starts with '/', it is used as a pattern.  If the argument
does not start with '/', it is taken literally, as a full tag name.